topology: fix some gcc10 warnings (labs, signess)

+++ src/topology/ctl.c
@@ -1363,7 +1363,7 @@ int tplg_decode_control_enum1(snd_tplg_t *tplg,
 		et->texts = tplg_calloc(heap, sizeof(char *) * ec->items);
 		if (!et->texts)
 			return -ENOMEM;
-		for (i = 0; i < ec->items; i++)
+		for (i = 0; (unsigned int)i < ec->items; i++)
 			et->texts[i] = ec->texts[i];
 	}
